Refactor comments controller

+++ app/controllers/comments_controller.rb
@@ -1,8 +1,7 @@
 class CommentsController < ApplicationController
   def create
     session[:return_to] ||= request.referer
-    @comment = commentable.comments.build(author_id: current_user.id, content: comment_params[:content])
-    @parent_class = @comment.commentable.class.name.downcase
+    @comment = current_user.comments.build(comment_params)
 
     respond_to do |format|
       if @comment.save
@@ -19,26 +18,17 @@ class CommentsController < ApplicationController
     session[:return_to] ||= request.referer
     @comment = current_user.comments.find(params[:id])
 
-    @comment.destroy
-
-    respond_to do |format|
-      format.html { redirect_to session.delete(:return_to), notice: "Comment deleted" }
-      format.js
+    if @comment.destroy
+      respond_to do |format|
+        format.html { redirect_to session.delete(:return_to), notice: "Comment deleted" }
+        format.js
+      end
     end
   end
 
   private
 
-  def parent_id
-    "#{params[:commentable].to_s}_id".downcase.to_sym
-  end
-
-  def commentable
-    # use the params to find the appropriate model and object
-    params[:commentable].capitalize.constantize.find(params[parent_id])
-  end
-
   def comment_params
-    params.require(:comment).permit(:content)
+    params.require(:comment).permit(:content, :commentable_type, :commentable_id)
   end
 end
